+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/src/OpenTelemetryBuilderExtensions.cs
@@ -122,8 +122,6 @@ namespace Azure.Monitor.OpenTelemetry.AspNetCore
                             .AddAzureMonitorTraceExporter());
 
             builder.WithMetrics(b => b
-                            .AddAspNetCoreInstrumentation()
-                            .AddHttpClientInstrumentation()
                             .AddAzureMonitorMetricExporter());
 
             builder.Services.AddLogging(logging =>